Extend 8SVX search to 512 first bytes
Either some samples are really badly formated, or I miss some
information as to why the header is so far in the file.

Note: really few files need this, and fewer have their header even
farther in the file. Those files are not super interesting anyway,
better to load them as raw and slice them by hand.